Move UserProfiles to the correct entity folder. Remove unneeded fields, add summary and full text to both
